move some util to site-utils
